refactor: Add API v1 namespace
