+++ config/routes.rb
@@ -5,8 +5,13 @@ Rails.application.routes.draw do
   # Can be used by load balancers and uptime monitors to verify that the app is live.
   get "up" => "rails/health#show", as: :rails_health_check
 
-  # Authentication
-  post "auth/google", to: "auth#google"
+  # API v1
+  namespace :api do
+    namespace :v1 do
+      # Authentication
+      post "auth/google", to: "auth#google"
+    end
+  end
 
   # Defines the root path route ("/")
   # root "posts#index"
